Integrated David Weatherall's calendar queue.
Added a simple, completely contrived script to generate
SRM-like sample paths to illustrate SRM in EECS colloquium
(see tcl/ex/srm-demo.tcl).

Fixed bug in Packet::copy.  "*p = *this" was bogus because
bits are now allocated in a separate buffer and this was
smashing the current packet's bits buffer to point to the
old one, which was both leaking memory and causing unwanted
sharing of private buffers.