Changes to multicast:
	1.  In the case of multiple mcast protocols running at the
same node, mrtObject_ keeps an array of protocols indexed by iif, so
in the case of upcalls like handle-cache-miss or handle-wrong-iif only
the protocols who owns the iif is called.

        2.  For the upcalls originated from classifier-mcast, returned
value 1 means that we want it to run classify on the packet again, 0
means that that's it, no more classifications needed.

        3.  CtrMcast instproc init now conforms all other
McastProto-derived init.

	4.  CHANGED SEMANTICS FOR SHARED-TREE <*,G> MFC-ENTRIES.  USE
'X' INSTEAD OF '*' IN add-mfc AND getReps.  Done this because of the
special meaning of '*' for TCL (wildcard), so instead of using ugly
"\\*", we use now 'x'.  ST.tcl is corrected to use the new semantics.

        5.  Finally, removed tcl/ex/newmcast/mcast2.tcl, because it
was using dynamicDM mode which ceased to exist as a separate protocol.