Adding changes for changes in the functionality of QuickStart. This includes the following:
A new default mapping, and alternate mapping, for the rate_function
encoding the Rate Request in the QuickStart packet:
* 1: Bps = rate * 10 KB;
* 2: Bps = 2 ^ rate * 1024;

Changes to TCP to respond correctly to the loss of a QuickStart packet.
The variable tcp_qs_recovery_ helps specify the response.

A new validation test pkt_drops to check the response to a packet drop.

The code is from Pasi Sarolahti.