Wpan patch from Iyappan Ramachandran (iyappan.ram@gmail.com)
1. Ability to put a WPAN node to sleep has been added with appropriate
actions to decrement energy in sleep state and in transitions. Appropriate
code has been added to enable node sleep at any time that the user wishes
from the tcl interface.
2. The earlier implementation was indefinitely attempting transimission of
a packet. This has been fixed; after backing of macMaxCSMABackoffs, a node
just reports access failure to the upper layer.
3. Owing to the lack of a detailed-enough PHY layer, CCA has to be done at
one instant (the end of the 8th symbol, in the earlier implementation)
instead of integration over 8 symbols. As this was leading to unrealistic
behavior, CCA has been modified to be done at the end of the 4th symbol but
reported to MAC at the end of the 8th.
4. Code has been added to decrement appropritate energy in the carrier
sensing process.
5. The WirelessChannel::sendUp() function call by Phy802 15 4::PD DATA
indication() causes receive energy to be decremented twice and has been
fixed.
6. Several other smaller bugs have been fixed.